backend/requirements.txt
asgiref==3.3.2
Django==3.2.0
django-filter==23.5
djangorestframework==3.12.4
django-debug-toolbar
gunicorn==20.0.4
djoser
pillow
psycopg2-binary~=2.8.6
python-dotenv
pytz==2020.1
sqlparse==0.3.1
requests==2.26.0